Smooth Spinners: Makes the loading "spinner" animation smoother in non-holo apps
The loading "spinner" animation is quite jerky by default when the progress xml's from
drawable are called in applications that don't use the holo progress animation. The holo
"spinner" animation is already smooth by default. This makes every other "spinner"
animation smooth.

We've been changing this value in all ROM's we release, both AOSP and CM based ones
as well as stock ROM's, since it gives a more polished look to the entire system. In the past
3 years we've been using this small tweak we have not noticed anything that would
adversely affect battery life or system operations as a whole.

Change-Id: I06573dd36130d14d8c6c99cda01540f1ec78bd95
Signed-off-by: Sar Castillo <sar.castillo@gmail.com>
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
Signed-off-by: RakeshBatra <raakesh.batra@rediffmail.com>